LPS-80165 Move logic added by LPS-29823 and LPS-65314 from VerifyWorkflow to an upgrade process
